+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog for NeoFS Node
 
 ### Added
 - Embedded Neo contracts in `contracts` dir (#2391)
+- `dump-names` command for adm
 
 ### Fixed